refactor: add examples for scc algorithms and reorganize into dedicated module (#830)
tasks:
- [x] Add examples to kosaraju_scc docs.
- [x] Add examples to tarjan_scc docs.
- [x] Move both algorithms into separate module (and re-export them).

---------
