+++ src/visit/traversal.rs
@@ -312,7 +312,7 @@ where
 ///
 /// **Note** that `Topo` only visits nodes that are not part of cycles,
 /// i.e. nodes in a true DAG. Use other visitors like [`DfsPostOrder`] or
-/// algorithms like [`kosaraju_scc`][crate::algo::kosaraju_scc] to handle
+/// algorithms like [`kosaraju_scc`][crate::algo::kosaraju_scc()] to handle
 /// graphs with possible cycles.
 #[derive(Clone)]
 pub struct Topo<N, VM> {
